- Bug fix    null pointer exception at image deletion, in onCreate method, when resumed under particular conditions.

+++ app/src/main/java/com/example/utente/calcolaorauscita/MainActivity.java
@@ -455,7 +455,7 @@ public class MainActivity extends ActionBarActivity {
 
         // Cancello il file quando ritorno dallo share picker
         if(requestCode==SHARE_PICKER){
-            if (imagePath.exists()) {
+            if (imagePath!=null && imagePath.exists()) {
                 imagePath.delete();
             }
         }
